put all park/unpark logic together, instead of OOP style
